Consolidate pipeline contracts in abstractions
