+++ src/FAI.Core/Abstractions.cs
@@ -1,25 +1,48 @@
+using System.Diagnostics.CodeAnalysis;
+
 // ReSharper disable once CheckNamespace
-namespace FAI.Core.Abstractions;
-
-/// <summary>
-/// Defines application-level inference operations.
-/// </summary>
-/// <typeparam name="TInput">The input type.</typeparam>
-/// <typeparam name="TOutput">The output type.</typeparam>
-public interface IInference<TInput, TOutput>
+namespace FAI.Core.Abstractions
 {
-    /// <summary>
-    /// Predicts one output for one input.
-    /// </summary>
-    Task<TOutput> Predict(TInput input);
 
     /// <summary>
-    /// Predicts outputs for a batch of inputs.
+    /// Defines application-level inference operations.
     /// </summary>
-    Task<TOutput[]> BatchPredict(ReadOnlyMemory<TInput> input);
+    /// <typeparam name="TInput">The input type.</typeparam>
+    /// <typeparam name="TOutput">The output type.</typeparam>
+    public interface IInference<TInput, TOutput>
+    {
+        /// <summary>
+        /// Predicts one output for one input.
+        /// </summary>
+        Task<TOutput> Predict(TInput input);
 
-    /// <summary>
-    /// Predicts outputs into a caller-provided buffer.
-    /// </summary>
-    Task BatchPredict(ReadOnlyMemory<TInput> input, Memory<TOutput> output);
+        /// <summary>
+        /// Predicts outputs for a batch of inputs.
+        /// </summary>
+        Task<TOutput[]> BatchPredict(ReadOnlyMemory<TInput> input);
+
+        /// <summary>
+        /// Predicts outputs into a caller-provided buffer.
+        /// </summary>
+        Task BatchPredict(ReadOnlyMemory<TInput> input, Memory<TOutput> output);
+    }
+}
+
+namespace FAI.Core.Pipelines
+{
+
+    public interface IPipeline<in TInput, TOutput>
+    {
+        ValueTask<TOutput> ExecuteAsync(TInput input, CancellationToken cancellationToken = default);
+    }
+
+    public interface IPreallocatingPipeline<in TInput, TOutput> : IPipeline<TInput, TOutput>
+    {
+        bool TryAllocateOutput(TInput input, [MaybeNullWhen(false)] out TOutput output);
+
+        ValueTask ExecuteAsync(
+            TInput input,
+            TOutput output,
+            CancellationToken cancellationToken = default);
+    }
 }
